requirements.txt
appdirs==1.4.3
asgiref==3.3.4
boto==2.49.0
boto3==1.17.62
botocore==1.20.62
Brotli==1.0.9
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.2
django-editorjs==0.2.1
django-modeltranslation==0.17.1
django-quill-editor==0.1.22
django-robots==4.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jmespath==0.10.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
s3transfer==0.4.2
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
